Decouple static val from static meth reflection
