+++ tests/run-with-compiler/tasty-interpreter.check
@@ -18,5 +18,5 @@ if
 
 55
 precompiledModule
-66
 55
+56
