+++ tests/run-with-compiler/tasty-interpreter/Precompiled.scala
@@ -4,10 +4,11 @@ trait Bar {
 }
 
 object Precompiled {
-  val staticVal  = 55
-  def staticMeth = 66
+  def staticMeth = 55
+  val staticVal  = 56
+
   // Todo
-  // def staticMeth1() = 66
+  // def staticMeth1() = 57
   // def staticMeth2(arg: Int) = arg
   // def staticMeth3(arg: Bar): Int = arg.meth()
